chore: update to Spring Boot to 3.1.12 and related dependencies.
- Update Spring Boot to version 3.1.12
- Update Spring dependency management plugin to 1.1.7
- Refresh OpenAPI documentation dependencies
- Minor adjustments to security configuration

+++ src/main/java/org/sitmun/domain/user_token/UserTokenRepository.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 @Tag(name = "token_user")
-public interface UserTokenRepository extends JpaRepository<UserToken, Long> {
+public interface UserTokenRepository extends JpaRepository<UserToken, Integer> {
   Optional<UserToken> findByUserMailAndTokenId(String userMail, String tokenId);
 
   Optional<UserToken> findByTokenId(String tokenId);
